Quiz owners

+++ src/app/quiz/actions.ts
@@ -2,10 +2,15 @@
 
 import prisma from '@/lib/db';
 import { revalidatePath } from 'next/cache';
-import { convertDbError, simulateLatency } from '@/lib/util';
+import { convertDbError, isUserAllowedToEdit } from '@/lib/util';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '@/lib/auth';
 
 export async function createQuiz(name: string, {}: {}) {
-  await simulateLatency();
+  let session = await getServerSession(authOptions);
+  if (!session) {
+    return { error: 'Unauthorized' };
+  }
   if (!name) {
     return { error: 'name is required' };
   }
@@ -13,6 +18,7 @@ export async function createQuiz(name: string, {}: {}) {
     await prisma.quiz.create({
       data: {
         name,
+        ownerId: session.user.id,
       },
     });
   } catch (e) {
@@ -22,6 +28,9 @@ export async function createQuiz(name: string, {}: {}) {
 }
 
 export async function deleteQuiz({ id }: { id: number }) {
+  if (!(await isUserAllowedToEdit(id))) {
+    return { error: 'Unauthorized' };
+  }
   try {
     await prisma.quiz.delete({
       where: { id },
